requirements-airflow.txt
# Airflow core dependencies
apache-airflow==2.7.1
apache-airflow-providers-http==4.7.0

# Additional dependencies for RAG integration
requests==2.31.0
python-dateutil==2.8.2

# Existing RAG dependencies (from requirements.txt)
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
openai==1.3.7
chromadb==0.4.18
sentence-transformers==2.2.2
pandas==2.1.4
python-docx==1.1.0
PyPDF2==3.0.1
python-pptx==0.6.23
openpyxl==3.1.2
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
atlassian-python-api==3.41.4
tableauserverclient==0.20
beautifulsoup4==4.12.2
apscheduler==3.10.4
python-dotenv==1.0.0
pydantic==2.5.0
loguru==0.7.2
tqdm==4.66.1
